Icon I made for PPM/2
This is an icon I made for PPM/2. I polled the staff of Ponyscape as well as two other servers and asked which one they liked the best, including all old icons, UnkN's icon, and DBot's new icon. This was the one that was chosen (there was a non-transparent version).